feat: get all the file extension (i.e file type) in a dir and its sub dir
Introduced a feature that gets all the file extension present in dir and
its subdir.
Useful to find all the type of files that can be sorted under a
category.

+++ src/sortium/file_utils.py
@@ -1,12 +1,15 @@
+import shutil
 from pathlib import Path
 from datetime import datetime
-import shutil
+from typing import List, Set
+
 
 class FileUtils:
-    '''
+    """
     FileUtils class for file utilities that provides various methods for working with files and directories and also are used in the Sorter class.
     A Custom FileUtils class can be provided to the Sorter class to satisfy the specific requirements.
-    '''
+    """
+
     def __init__(self) -> None:
         """
         Initializes an instance of the FileUtils class.
@@ -15,8 +18,26 @@ class FileUtils:
         upon instantiation.
         """
         pass
+    
+    def _get_files_and_sub_dir(
+        self,
+        folder_path: str,
+        ignore_dir: list[str] | None = None,
+    ) -> tuple[List[str], List[str]]:
+        source_root: Path = Path(folder_path)
+        # Get name of the sub directories ignoring the one in ignore_dir list.
+        sub_dir_list = self.get_subdirectories_names(str(source_root), ignore_dir)
 
-    def get_file_modified_date(self,file_path: str) -> datetime:
+        # Get the list of files to be moved.
+        file_list = [
+            item.name
+            for item in source_root.iterdir()
+            if item.is_file() and item.name not in (ignore_dir or [])
+        ]
+
+        return sub_dir_list, file_list
+    
+    def get_file_modified_date(self, file_path: str) -> datetime:
         """
         Returns the last modified datetime of a file.
 
@@ -34,9 +55,8 @@ class FileUtils:
             raise FileNotFoundError(f"File does not exist: {file_path}")
         return datetime.fromtimestamp(path.stat().st_mtime)
 
-
-    def get_subdirectories_names(self,
-        folder_path: str, ignore_dir: list[str] = None
+    def get_subdirectories_names(
+        self, folder_path: str, ignore_dir: list[str] = None
     ) -> list[str]:
         """
         Returns a list of subdirectory names in a given folder, excluding any specified to be ignored.
@@ -58,11 +78,11 @@ class FileUtils:
         ]
         return sub_dir_list
 
-
-    def flatten_dir(self,
+    def flatten_dir(
+        self,
         folder_path: str,
         dest_folder_path: str,
-        ignore_dir: list[str] = None,
+        ignore_dir: list[str] | None = None,
         rm_subdir: bool = False,
     ) -> None:
         """
@@ -75,8 +95,7 @@ class FileUtils:
             folder_path (str): Path to the root folder containing subdirectories with files.
             dest_folder_path (str): Path to the folder where all files should be moved.
             ignore_dir (list[str]): Names of subdirectories within `folder_path` that should be ignored during processing.
-            rm_subdir (bool): If True, subdirectories will be removed after moving their contents.
-
+            rm_subdir (bool): If True, subdirectories will be removed after moving their contents. Default is False.
         Raises:
             FileNotFoundError: If the root folder (`folder_path`) does not exist.
 
@@ -93,15 +112,10 @@ class FileUtils:
         dest_root.mkdir(parents=True, exist_ok=True)
 
         try:
-            # Get name of the sub directories ignoring the one in ignore_dir list.
-            sub_dir_list = self.get_subdirectories_names(str(source_root), ignore_dir)
-
-            # Get the list of files to be moved.
-            file_list = [
-                item.name
-                for item in source_root.iterdir()
-                if item.is_file() and item.name not in (ignore_dir or [])
-            ]
+            # Get the list of files and sub directories.
+            sub_dir_list, file_list = self._get_files_and_sub_dir(
+                folder_path, ignore_dir
+            )
 
             # If file_list empty then then return the function for sub_dir.
             if sub_dir_list and not file_list:
@@ -130,17 +144,60 @@ class FileUtils:
                         )
 
                     # Remove the sub directories if rm_subdir is True.
-                    if rm_subdir:
-                        for sub_dir_name in sub_dir_list:
-                            sub_dir_path = source_root / sub_dir_name
-                            try:
-                                shutil.rmtree(sub_dir_path)
-                            except Exception as e:
-                                print(f"Failed to remove directory '{sub_dir_path}': {e}")
+            if rm_subdir:
+                for sub_dir_name in sub_dir_list:
+                    sub_dir_path = source_root / sub_dir_name
+                    try:
+                        if sub_dir_path != dest_root:
+                            shutil.rmtree(sub_dir_path)
+                    except Exception as e:
+                        print(
+                            f"Failed to remove directory '{sub_dir_path}': {e}"
+                        )   
 
         except Exception as e:
             print(f"Error occurred while cleaning up folders: {e}")
 
-    
+    def find_unique_extensions(
+        self, source_path: str, ignore_dir: list[str] | None = None
+    ) -> Set[str]:
+        """
+        Recursively finds all unique file extensions in a given directory and its subdirectories.
 
+        Args:
+            source_path (str): Path to the root directory.
+            ignore_dir (list[str], optional): List of directory names to ignore. Defaults to None.
 
+        Returns:
+            Set[str]: A set of unique file extensions found in the directory tree.
+
+        Raises:
+            FileNotFoundError: If the source_path does not exist.
+        """
+        source_root: Path = Path(source_path)
+        if not source_root.exists():
+            raise FileNotFoundError(
+                f"The folder path '{str(source_root)}' does not exist."
+            )
+
+        extension_list: Set[str] = set()
+
+        try:
+            sub_dir_list, file_list = self._get_files_and_sub_dir(
+                str(source_root), ignore_dir
+            )
+
+            for name in file_list:
+                source_item = source_root / name
+                extension_list.add(source_item.suffix.lower())
+
+            for sub_dir_name in sub_dir_list:
+                sub_dir_path = source_root / sub_dir_name
+                extension_list.update(
+                    self.find_unique_extensions(str(sub_dir_path), ignore_dir)
+                )
+
+        except Exception as e:
+            print(f"Error occurred while finding unique extensions: {e}")
+
+        return extension_list
